refactor(storage): extract envInt into core/env helper
Per review: the generic env-var parser is not pool-specific, move it
out of resizePool into core/env.ts.
